# local-hobbies-backend/openapi.yaml
openapi: 3.0.3
info:
  title: Local Hobbies API
  description: API for a user-matching application based on hobbies, location, and availability.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://api.example.com/api
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /auth/signup:
    post:
      tags:
        - Auth
      summary: Create a new user account
      description: Register a new user and receive JWT token with user information
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthSignupRequest'
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/login:
    post:
      tags:
        - Auth
      summary: Log into an existing account
      description: Authenticate user and receive JWT token with user information
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==================== Users ====================
...
  /hobbies:
    get:
      tags:
        - Hobbies
      summary: Get current user's hobbies
      responses:
        '200':
          description: Hobbies retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HobbyResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Hobbies
      summary: Add a hobby to current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HobbyCreationRequest'
      responses:
        '201':
          description: Hobby added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HobbyResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /hobbies/{hobbyId}:
    put:
      tags:
        - Hobbies
      summary: Update a hobby
      description: Update the experience level for one of the current user's hobbies
      parameters:
        - name: hobbyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HobbyUpdateRequest'
      responses:
        '200':
          description: Hobby experience level updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HobbyResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Hobbies
      summary: Remove a hobby from current user's profile
      parameters:
        - name: hobbyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Hobby deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /hobbies/global:
    get:
      tags:
        - Hobbies
      summary: Get global hobbies
      description: Get list of all hobbies users can select from. Hobbies are organized first by category, then by name.
      security: []
      responses:
        '200':
          description: Global hobbies retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GlobalHobbyResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /hobbies/{hobbyId}/photos/upload-url:
    post:
      tags:
        - Hobbies
      summary: Request hobby photo upload URL
      description: Request a presigned URL to upload a hobby photo to the cloud
      parameters:
        - name: hobbyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned URL received successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /hobbies/photos:
    get:
      tags:
        - Hobbies
      summary: Get current user's hobby photos
      responses:
        '200':
          description: Hobby photos retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HobbyPhotoResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /hobbies/photos/{photoId}:
    put:
      tags:
        - Hobbies
      summary: Update hobby photo
      description: Update the hobby tag or the caption of a hobby photo
      parameters:
        - name: photoId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HobbyPhotoUpdateRequest'
      responses:
        '200':
          description: Hobby photo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HobbyPhotoResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Hobbies
      summary: Delete a user's hobby photo
      parameters:
        - name: photoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Hobby photo deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==================== Availabilities ====================
...
  /matches/search:
    post:
      tags:
        - Matches
      summary: Search for matching users
      description: Find users based on hobbies, location, availability, and preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchSearchRequest'
      responses:
        '200':
          description: Search completed successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchSearchResultResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /matches/saved:
    get:
      tags:
        - Matches
      summary: Get current user's saved matches
      description: Get saved matches organized by hobby category, then hobby, then save date
      responses:
        '200':
          description: Saved matches retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SavedMatchResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Matches
      summary: Save a user as a match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavedMatchCreationRequest'
      responses:
        '201':
          description: Match saved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedMatchResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /matches/saved/{matchId}:
    put:
      tags:
        - Matches
      summary: Update saved match
      description: Update the notes on a saved match
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavedMatchUpdateRequest'
      responses:
        '200':
          description: Match updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedMatchResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Matches
      summary: Delete a saved match
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Match deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /matches/deleted:
    get:
      tags:
        - Matches
      summary: Get current user's deleted saved matches
      responses:
        '200':
          description: Deleted saved matches retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SavedMatchResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /matches/deleted/{matchId}:
    post:
      tags:
        - Matches
      summary: Restore a recently-deleted Match
      description: Restore a match that was deleted in the last 30 days
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: "Match restored successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedMatchResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'

# ==================== Components ====================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login, /auth/signup, or /users/onboarding
  schemas:
...
    AuthSignupRequest:
      type: object
      required:
        - username
        - password
        - email
      properties:
        username:
          type: string
          example: alice_smith
        password:
          type: string
          example: ExamplePassword12#
        email:
          type: string
          example: alice@example.com
    AuthLoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: alice_smith
        password:
          type: string
          example: ExamplePassword12#
    AuthResponse:
      type: object
      required:
        - auth
        - user
      description: response to successful signup, login, and onboarding
      properties:
        auth:
          type: object
          required:
            - accesToken
            - tokenType
            - expirationTime
            - refreshToken
          properties:
            accessToken:
              type: string
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
            tokenType:
              type: string
              example: Bearer
            expirationTime:
              type: string
              format: date-time
              description: ISO 8601 UTC timestamp
              example: "2026-02-08T12:00:00Z"
            refreshToken:
              type: string
              example: def502003a4b8c9d7e6f5a4b3c2d1e0f...
        user:
          type: object
          required:
            - id
            - onboardingComplete
          properties:
            id:
              type: integer
              example: 123
            onboardingComplete:
              type: boolean
              description: used to route to either homepage or onboarding page
              example: false

    # ==================== User, Homepage, & Profile Schemas ====================
...
    HobbyCreationRequest:
      type: object
      required:
        - name
        - experienceLevel
      properties:
        name:
          type: string
          example: Hiking
        experienceLevel:
          type: string
          enum: [Beginner, Intermediate, Advanced]
          example: Intermediate
    HobbyUpdateRequest:
      type: object
      required:
        - experienceLevel
      properties:
        experienceLevel:
          type: string
          enum: [Beginner, Intermediate, Advanced]
          example: Advanced
    HobbyResponse:
      type: object
      required:
        - id
        - name
        - category
        - experienceLevel
      properties:
        id:
          type: integer
          example: 49
        name:
          type: string
          example: Hiking
        category:
          type: string
          example: Fitness
        experienceLevel:
          type: string
          enum: [Beginner, Intermediate, Advanced]
          example: Intermediate
...
    HobbyPhotoUpdateRequest:
      type: object
      description: only contains changed fields
      properties:
        hobbyId:
          type: integer
          description: should only be changed if current value is incorrect
          example: 49
        caption:
          type: string
          nullable: true
          example: Had a great time!
    HobbyPhotoResponse:
      type: object
      required:
        - id
        - hobbyId
        - photoUrl
        - caption
      properties:
        id:
          type: integer
          example: 335
        hobbyId:
          type: integer
          example: 49
        photoUrl:
          type: string
          example: https://example.com/photos/123/hobby135.jpg
        caption:
          type: string
          nullable: true
          example: Had a great time!
    GlobalHobbyResponse:
      type: object
      required:
        - name
        - category
      properties:
        name:
          type: string
          example: Hiking
        category:
          type: string
          example: Fitness
...
    AvailabilityOverlapResponse:
      type: object
      required:
        - distanceKilometers
        - start
        - end
      properties:
        distanceKilometers:
          type: number
          format: double
          description: distance between the two availabilities
          example: 4.3
        start:
          type: string
          format: date-time
          description: ISO 8601 UTC timestamp
          example: "2026-02-08T12:00:00Z"
        end:
          type: string
          format: date-time
          description: ISO 8601 UTC timestamp
          example: "2026-02-09T10:00:00Z"

    # ==================== Match Schemas ====================
    MatchedUser:
      type: object
      required:
        - id
        - name
        - profilePhotoUrl
        - lastSessionTime
      properties:
        id:
          type: integer
          example: 123
        name:
          type: string
          example: Alice Smith
        profilePhotoUrl:
          type: string
          nullable: true
          example: https://example.com/photos/123/profile.jpg
        lastSessionTime:
          type: string
          format: date-time
          description: ISO 8601 UTC timestamp
          example: "2026-02-08T12:00:00Z"
    MatchSearchRequest:
      type: object
      required:
        - hobby
        - radiusKilometers
        - minimumOverlapMinutes
        - filters
        - hardFilters
        - softFilters
      properties:
        hobby:
          type: string
          description: one of current user's hobbies
          example: Hiking
        radiusKilometers:
          type: integer
          minimum: 0
        minimumOverlapMinutes:
          type: integer
          minimum: 0
        filters:
          type: object
          properties:
            genders:
              type: array
              items:
                type: string
                enum: [Man, Woman, Nonbinary]
            minAge:
              type: integer
              minimum: 13
            maxAge:
              type: integer
              maximum: 120F
            experienceLevel:
              type: string
              enum: [Beginner, Intermediate, Advanced]
        hardFilters:
          type: array
          items:
            type: string
          description: filters that the user asserts must be satisfied; must be a subset of the filters that were passed; cannot overlap with soft filters
          example: ["genders", "minAge"]
        softFilters:
          type: array
          items:
            type: string
          description: filters used for prioritizing results; must be a subset of the filters that were passed; cannot overlap with soft filters
          example: ["experienceLevel"]
    MatchSearchResultResponse:
      type: object
      required:
        - matchedUser
        - hobby
        - distanceKilometers
        - overlappingAvailabilities
      properties:
        matchedUser:
          $ref: '#/components/schemas/MatchedUser'
        hobby:
          type: string
          example: Hiking
        distanceKilometers:
          type: number
          format: double
          example: 4.3
        overlappingAvailabilities:
          type: array
          items:
            $ref: '#/components/schemas/AvailabilityOverlapResponse'
    SavedMatchCreationRequest:
      type: object
      required:
        - savedUserId
        - hobby
      properties:
        savedUserId:
          type: integer
          example: 123
        hobby:
          type: string
          example: Hiking
        notes:
          type: string
          example: Brings her dog on hikes
    SavedMatchUpdateRequest:
      type: object
      required:
        - notes
      properties:
        notes:
          type: string
          example: Brings her dog Max on hikes
    SavedMatchResponse:
      type: object
      required:
        - id
        - userId
        - savedUser
        - hobby
        - notes
        - creationTime
      properties:
        id:
          type: integer
          example: 5
        userId:
          type: integer
          example: 80
        savedUser:
          $ref: '#/components/schemas/MatchedUser'
        hobby:
          type: string
          example: Hiking
        notes:
          type: string
          nullable: true
          example: Brings her dog on hikes
        creationTime:
          type: string
          format: date-time
          description: ISO 8601 UTC timestamp
          example: "2026-01-20T15:00:00Z"

    # ==================== Location Schemas ====================
...
    UploadUrlRequest:
      type: object
      description: request for presigned URL to upload file to cloud storage
      required:
        - fileName
        - contentType
      properties:
        fileName:
          type: string
          example: "vacation_hiking.jpg"
        contentType:
          type: string
          example: "image/jpeg"
    UploadUrlResponse:
      type: object
      required:
        - fileId
        - uploadUrl
        - expirationTime
      properties:
        fileId:
          type: string
          format: uuid
        uploadUrl:
          type: string
          format: uri
          description: cloud storage URL for image binary save request
        expirationTime:
          type: string
          format: date-time
          description: ISO 8601 UTC timestamp
          example: "2026-02-08T12:00:00Z"

    # ==================== Error Schemas ====================
    GlobalError:
      type: object
      required:
        - errorCode
        - errorMessage
        - timestamp
        - endpointURI
      properties:
        errorCode:
          type: string
          example: RESOURCE_NOT_FOUND
        errorMessage:
          type: string
          example: The requested hobby was not found.
        timestamp:
          type: string
          format: date-time
        endpointURI:
          type: string
          example: "/api/hobbies/Hiking"
  responses:
...
    BadRequestError: #400
      description: "The request was invalid or cannot be served."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GlobalError'
    UnauthorizedError: #401
      description: "Authentication is required or has failed."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GlobalError'
    ForbiddenError: #403
      description: "Insufficient permissions to access this resource."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GlobalError'
    NotFoundError: #404
      description: "The requested resource could not be found."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GlobalError'
    ConflictError: #409
      description: "The request could not be completed due to a conflict with the current state of the resource."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GlobalError'
    InternalServerError: #500
      description: "An unexpected error occurred on the server."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GlobalError'